abstracted out the player transformation to the playerservice file
