+++ server/src/Service/Simulation/SimulationService.php
@@ -41,8 +41,7 @@ class SimulationService
     public function simulateTournament(): iterable
     {
         //successfully converted all players to playersimobjects
-        $allActivePlayers = $this->playerService->getAllActivePlayers();
-        $allPlayerSimObjects = $this->getPlayerSimObjects($allActivePlayers);
+        $allPlayerSimObjects = $this->playerService->getActivePlayerSimObjects();
 
         //hard coded for now
         $currentCourse = $this->courseService->getCourseById(3);
@@ -55,24 +54,5 @@ class SimulationService
         $holeResults = $this->iterators->playerIterator($allPlayerSimObjects, $transformedHoles);
 
         return $holeResults;
-    }
-
-    public function getPlayerSimObjects($allActivePlayers): iterable
-    {
-        $FLOOR_CEILING = new \stdClass();
-        $FLOOR_CEILING->c1xFloorCeiling = [0.55, 0.92];
-        $FLOOR_CEILING->c2FloorCeiling = [0.01, 0.39];
-        $FLOOR_CEILING->parkedFloorCeiling = [0.01, 0.16];
-        $FLOOR_CEILING->c1RegFloorCeiling = [0.16, 0.46];
-        $FLOOR_CEILING->c2RegFloorCeiling = [0.29, 0.73];
-        $FLOOR_CEILING->scrambleFloorCeiling = [0.14, 0.64];
-
-        $allPlayersConverted = array();
-
-        foreach($allActivePlayers as $player) {
-            $allPlayersConverted[] = $this->playerIngester->convertPlayer($player,$FLOOR_CEILING);
-        }
-
-        return $allPlayersConverted;
     }
 }
